Created Components folder & added Stack components for the navigation, also installed libraries related to navigation
