+++ Components/Stack.js
@@ -1,0 +1,29 @@
+import React from 'react';
+import {View, Text, StyleSheet} from 'react-native';
+import {createStackNavigator} from '@react-navigation/stack';
+
+const Stack = createStackNavigator();
+
+const MyStack = () => {
+  return (
+    <Stack.Navigator>
+      <Stack.screen name="Home" component={Home}/>
+    </Stack.Navigator>
+    <View style={styles.container}>
+      <Text style={styles.text}>Hello world</Text>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  text: {
+    fontSize: 25,
+  },
+});
+
+export default MyStack;
